+++ demo/meeting/src/containers/DeviceSelection/MicSelection/AudioActivityPreview.tsx
@@ -1,3 +1,6 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 import React, { useEffect, useRef } from 'react';
 import { Label } from 'amazon-chime-sdk-component-library-react';
 import { useAudioVideo } from '../../../../../../src';