+++ demo/meeting/src/providers/ContentShareControlProvider.tsx
@@ -1,3 +1,6 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 import React, { useState, createContext, useCallback, useContext } from 'react';
 import { ContentShareControlContextType } from '../types';
 import { useAudioVideo } from '../../../../src';