+++ demo/meeting/src/providers/SIPMeetingProvider.tsx
@@ -1,3 +1,6 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 import React, { ReactNode } from 'react';
 import { useMeetingManager } from '../../../../src';
 import MeetingManager from '../../../../src/providers/MeetingProvider/MeetingManager';