+++ demo/meeting/src/utils/select-options-format.ts
@@ -1,3 +1,6 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 export type SelectOptionsType = {
   value: string | number;
   label: string;