+++ demo/meeting/src/views/index.tsx
@@ -1,2 +1,5 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 export { default as DeviceSetup } from './DeviceSetup';
 export { default as Home } from './Home';