+++ src/components/ControlBar/ControlBar.stories.tsx
@@ -1,3 +1,6 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 import React, { useState } from 'react';
 import { select, boolean } from '@storybook/addon-knobs';