+++ src/components/Portal/index.tsx
@@ -1,3 +1,6 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 import { FC, useEffect, useState } from 'react';
 import { createPortal } from 'react-dom';