+++ src/index.ts
@@ -1,3 +1,6 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 // Components
 export { Badge } from './components/Badge';
 export { Button } from './components/Button/';