+++ src/theme/globalStyles.ts
@@ -1,3 +1,6 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 import { createGlobalStyle } from 'styled-components';
 import reset from 'styled-reset';