+++ tst/components/Textarea/index.test.tsx
@@ -1,3 +1,6 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 import '@testing-library/jest-dom';
 import React from 'react';
 import userEvent from "@testing-library/user-event";