+++ tst/snapshots/Badge/index.test.tsx
@@ -1,3 +1,6 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 describe('Badge', () => {
     it('default status', async () => {
       await page.goto('http://host.docker.internal:9009/iframe.html?id=badge--basic-badge');