+++ tst/snapshots/Flex/index.test.tsx
@@ -1,3 +1,6 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 describe('Flex Snapshots', () => {
     it('fill space centered', async () => {
       await page.goto('http://host.docker.internal:9009/iframe.html?id=flex--fill-space-centered');