+++ tst/utils/trap-focus.test.tsx
@@ -1,3 +1,6 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 import React from 'react';
 import { render } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";